Time: 0 ms (100%), Space: 13.3 MB (76.18%) - LeetHub

+++ 0074-search-a-2d-matrix/0074-search-a-2d-matrix.cpp
@@ -1,3 +1,7 @@
+// Time Complexity: O(log(NxM)), where N = given row number, M = given column number.
+// Reason: We are applying binary search on the imaginary 1D array of size NxM.
+
+
 class Solution {
 public:
     bool searchMatrix(vector<vector<int>>& matrix, int target) {
@@ -6,7 +10,7 @@ public:
      int m=matrix[0].size();//no of columns in row
      int high=n*m-1;
      while(low<=high){
-        long long int mid=low+(high-low)/2;
+      int mid=low+(high-low)/2;
         int row=mid/m;
         int col=mid%m;
         if(matrix[row][col]==target){
